# local-api-docs.yaml
openapi: 3.0.1
info:
  title: gateways-api
  description: a private microservice that serves as gateways and devices
  version: v1.0.0
servers:
  - url: http://localhost:8080/gateways-api
    description: Generated server url
security:
  - basicAuth: []
paths:
...
  /api/rest/gateway/{gateway_uuid}:
    get:
      tags:
        - getGateway
      summary: get a single Gateway resource
      operationId: getGateway
      parameters:
        - name: gateway_uuid
          in: path
          description: Uuid of the wanted gateway. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3c221f5-3c2d-11ec-a662-0242ac160003
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad gateway_uuid, not in uuid format.:
                  description: "An example Response to a bad gateway_uuid, not in\
                    \ uuid format."
                  value:
                    error: "Validation failed, getGateway.gateway_uuid: Requested\
                      \ uuid q3c23161-3c2d-11ec-a662-0242ac160003 is not valid."
        "404":
          description: Requested resource is not found
          content:
            application/json:
              examples:
                An example Response to gateway_uuid not found.:
                  description: An example Response to gateway_uuid not found.
                  value:
                    error: Requested Gateway with uuid ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                      is not Found
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:45:26.377+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/gateway/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
        "200":
          description: successfully retrieved a single Gateway resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayModel'
              examples:
                An example Response to getGatewayPage endpoint.:
                  description: An example Response to getGatewayPage endpoint.
                  value:
                    gatewayUuid: a3c23161-3c2d-11ec-a662-0242ac160003
                    gatewayName: gateway-a3c2316a-3c2d-11ec-a662-0242ac160003
                    gatewayIpv4: 70.22.2.45
    put:
      tags:
        - updateGateway
      summary: Update Gateway details
      operationId: updateGateway
      parameters:
        - name: gateway_uuid
          in: path
          description: Uuid of the gateway to be updated. Cannot be empty.
          required: true
          schema:
            type: string
          example: a3c221f5-3c2d-11ec-a662-0242ac160003
      requestBody:
        description: new values of the Gateway to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbstractGateway'
            example:
              gatewayUuid: a3c231e0-3c2d-11ec-a662-0242ac160003
              gatewayName: gateway-a3c231e8-3c2d-11ec-a662-0242ac160003
              gatewayIpv4: 17.50.0.52
        required: true
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad gateway_uuid, not in uuid format.:
                  description: "An example Response to a bad gateway_uuid, not in\
                    \ uuid format."
                  value:
                    error: "Validation failed, updateGateway.gateway_uuid: Requested\
                      \ uuid q3c23161-3c2d-11ec-a662-0242ac160003 is not valid."
        "404":
          description: Requested resource is not found
          content:
            application/json:
              examples:
                An example Response to gateway_uuid not found.:
                  description: An example Response to gateway_uuid not found.
                  value:
                    error: Requested Gateway with uuid ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
                      is not Found
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:45:26.377+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/gateway/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
        "200":
          description: successfully updated a single Gateway entity
          content:
            application/json:
              examples:
                An example Response to updateGateway endpoint.:
                  description: An example Response to updateGateway endpoint.
                  value:
                    message: Gateway updated successfully
...
  /api/rest/gateway:
    post:
      tags:
        - createGateway
      summary: create a Gateway
      operationId: createGateway
      requestBody:
        description: Gateway to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbstractGateway'
            example:
              gatewayUuid: a3c231e0-3c2d-11ec-a662-0242ac160003
              gatewayName: gateway-a3c231e8-3c2d-11ec-a662-0242ac160003
              gatewayIpv4: 17.50.0.52
        required: true
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad request body.:
                  description: An example Response to a bad request body.
                  value:
                    error: "Validation failed, failed to bind value null to field\
                      \ gatewayUuid, failed to bind value null to field gatewayIpv4,\
                      \ failed to bind value null to field gatewayName."
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:45:26.377+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/gateway/ad1275e3-b3ef-48e0-8ce8-fd3fbdf0080f
        "200":
          description: successfully created Gateway entity
          content:
            application/json:
              examples:
                An example Response to createGateway endpoint.:
                  description: An example Response to createGateway endpoint.
                  value:
                    message: Gateway created successfully
  /api/rest/gateways:
    get:
      tags:
        - getGatewayPage
      summary: get a page of Gateways
      operationId: getGatewayPage
      parameters:
        - name: gatewayPaginationRequest
          in: query
          description: "object contains the Query parameters size and page, default\
          \ values for size is 10 and for page is 0. Cannot be empty."
          required: true
          schema:
            $ref: '#/components/schemas/GatewayPaginationRequest'
          example:
            page: 0
            size: 10
      responses:
        "400":
          description: "bad request, not valid parameter"
          content:
            application/json:
              examples:
                An example Response to a bad page and bad size values.:
                  description: An example Response to a bad page and bad size values.
                  value:
                    error: "Validation failed, failed to bind value four to field\
                      \ page, failed to bind value ten to field size."
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              examples:
                An example Response to an Internal Server Error.:
                  description: An example Response to an Internal Server Error.
                  value:
                    error: something went wrong.
        "401":
          description: unAuthorized
          content:
            application/json:
              examples:
                An example Response to an un-Authorized request.:
                  description: An example Response to an un-Authorized request.
                  value:
                    timestamp: 2021-11-16T03:45:26.377+00:00
                    status: 401
                    error: Unauthorized
                    path: /gateways-api/api/rest/gateways
        "200":
          description: successfully retrieved a single page of Gateways
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbstractGateway'
              examples:
                An example Response to getGatewayPage endpoint.:
                  description: An example Response to getGatewayPage endpoint.
                  value:
                    - gatewayUuid: a3c23161-3c2d-11ec-a662-0242ac160003
                      gatewayName: gateway-a3c2316a-3c2d-11ec-a662-0242ac160003
                      gatewayIpv4: 70.22.2.45
                    - gatewayUuid: a3c231e0-3c2d-11ec-a662-0242ac160003
                      gatewayName: gateway-a3c231e8-3c2d-11ec-a662-0242ac160003
                      gatewayIpv4: 17.50.0.52
components:
  schemas:
    DeviceModel:
      type: object
      properties:
        devicesUuid:
          type: string
        devicesName:
          type: string
        vendor:
          type: string
        deviceCreationDate:
          type: string
          format: date-time
        status:
          type: boolean
    AbstractGateway:
      type: object
      properties:
        gatewayUuid:
          type: string
        gatewayName:
          type: string
        gatewayIpv4:
          type: string
    GatewayPaginationRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    GatewayModel:
      type: object
      properties:
        gatewayUuid:
          type: string
        gatewayName:
          type: string
        gatewayIpv4:
          type: string
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceModel'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic